+++ components/layout/Nav.js
@@ -1,5 +1,4 @@
 import { useState, useEffect, useRef } from 'react';
-import Image from 'next/legacy/image';
 import { useRouter } from 'next/router';
 import Container from '@/components/containers/Container';
 import styles from '@/styles/Nav.module.scss';
@@ -55,17 +54,14 @@ export default function Nav() {
       <Container>
         <div ref={containerRef}>
           <nav className={styles.nav}>
-            <div className={styles.nav__logo}>
-              <a href='/' passhref='true'>
-                <Image
-                  src='/images/svg/logo.svg'
-                  height={115}
-                  width={180}
-                  alt='Logo'
-                  title='Go to the Homepage'
-                />
-              </a>
-            </div>
+            <a href='/' passhref='true'>
+              <img
+                className={styles.nav__logo}
+                src='/images/svg/logo.svg'
+                alt='Logo'
+                title='Go to the Homepage'
+              />
+            </a>
             <ul
               className={`${styles.nav__links} ${active ? styles.active : ''}`}
             >